refactor: 방명록 공간 컴포넌트 분리
