+++ components/three/index.ts
@@ -1,0 +1,10 @@
+import Desk from '@/components/three/Desk';
+import Floor from '@/components/three/Floor';
+import Keyboard from '@/components/three/Keyboard';
+import Keycap from '@/components/three/Keycap';
+import Laptop from '@/components/three/Laptop';
+import Monitor from '@/components/three/Monitor';
+import Scene from '@/components/three/Scene';
+import Wall from '@/components/three/Wall';
+
+export { Desk, Floor, Keyboard, Keycap, Laptop, Monitor, Scene, Wall };
